requirements.txt
gradio==4.44.1
openai
langchain-community
weaviate-client
python-dotenv
grpcio==1.60.1